cookbook/pocoflow-rag/requirements.txt
pocoflow>=0.2.0
anthropic>=0.25.0
openai>=1.0.0
python-dotenv>=1.0
numpy>=1.20.0
faiss-cpu>=1.7.0
